+++ qoqo_qiskit/src/qoqo_qiskit/backend/post_processing.py
@@ -22,16 +22,16 @@ from ..models import RegistersWithLengths
 
 
 def _counts_to_registers(
-    counts: Any, mem: bool, clas_regs_lengths: Dict[str, int]
+    counts: Any, mem: bool, bit_regs_lengths: Dict[str, int]
 ) -> List[List[List[bool]]]:
     bit_map: List[List[List[bool]]] = []
     reg_num = 0
-    for key in clas_regs_lengths:
-        reg_num += clas_regs_lengths[key]
+    for key in bit_regs_lengths:
+        reg_num += bit_regs_lengths[key]
     for _ in range(reg_num):
         bit_map.append([])
     for key in counts:
-        splitted = _split(key, clas_regs_lengths)
+        splitted = _split(key, bit_regs_lengths)
         for i, measurement in enumerate(splitted):
             transf_measurement = _bit_to_bool(measurement)
             if mem:
@@ -56,17 +56,17 @@ def _bit_to_bool(element: str) -> List[bool]:
     return ret
 
 
-def _split(element: str, clas_regs_lengths: Dict[str, int]) -> List[str]:
+def _split(element: str, bit_regs_lengths: Dict[str, int]) -> List[str]:
     splitted: list[str] = []
     if " " in element:
         splitted = element.split()
         splitted.reverse()
     else:
         element = element[::-1]
-        for key in clas_regs_lengths:
-            splitted.append(element[: clas_regs_lengths[key] :])
+        for key in bit_regs_lengths:
+            splitted.append(element[: bit_regs_lengths[key] :])
             splitted[-1] = splitted[-1][::-1]
-            element = element[clas_regs_lengths[key] :]
+            element = element[bit_regs_lengths[key] :]
     return splitted
 
 
@@ -74,7 +74,8 @@ def _transform_job_result_single(
     memory: bool,
     sim_type: str,
     result: Result,
-    output_registers: RegistersWithLengths,
+    output_register: RegistersWithLengths,
+    input_bit_circuit: Optional[Circuit],
     res_index: Optional[int] = 0,
 ) -> None:
     # res_index is the index of the chosen result to extract in case
@@ -83,20 +84,24 @@ def _transform_job_result_single(
         transformed_counts = _counts_to_registers(
             result.get_memory(res_index) if memory else result.get_counts(res_index),
             memory,
-            output_registers.clas_regs_lengths,
+            output_register.bit_regs_lengths,
         )
-        for i, reg in enumerate(output_registers.registers.bit_register_dict):
-            output_registers.registers.bit_register_dict[reg] = [
+        for i, reg in enumerate(output_register.registers.bit_register_dict):
+            output_register.registers.bit_register_dict[reg] = [
                 shot[::-1] for shot in transformed_counts[i]
             ]
+        if input_bit_circuit:
+            for input_bit_op in input_bit_circuit:
+                for bit_result in output_register.registers.bit_register_dict[input_bit_op.name()]:
+                    bit_result[input_bit_op.index()] = input_bit_op.value()
     elif sim_type == "statevector":
         vector = list(np.asarray(result.data(res_index)["statevector"]).flatten())
-        for reg in output_registers.registers.complex_register_dict:
-            output_registers.registers.complex_register_dict[reg].append(vector)
+        for reg in output_register.registers.complex_register_dict:
+            output_register.registers.complex_register_dict[reg].append(vector)
     elif sim_type == "density_matrix":
         vector = list(np.asarray(result.data(res_index)["density_matrix"]).flatten())
-        for reg in output_registers.registers.complex_register_dict:
-            output_registers.registers.complex_register_dict[reg].append(vector)
+        for reg in output_register.registers.complex_register_dict:
+            output_register.registers.complex_register_dict[reg].append(vector)
 
 
 def _transform_job_result_list(
@@ -104,15 +109,20 @@ def _transform_job_result_list(
     sim_type: str,
     result: Result,
     output_registers: List[RegistersWithLengths],
+    input_bit_circuits: List[Optional[Circuit]],
 ) -> None:
     if sim_type == "automatic":
         res_list = result.get_memory() if memory else result.get_counts()
-        for res, regs in zip(res_list, output_registers):
-            transformed_counts = _counts_to_registers(res, memory, regs.clas_regs_lengths)
+        for res, regs, input_bit_circuit in zip(res_list, output_registers, input_bit_circuits):
+            transformed_counts = _counts_to_registers(res, memory, regs.bit_regs_lengths)
             for i, reg in enumerate(regs.registers.bit_register_dict):
                 regs.registers.bit_register_dict[reg] = [
                     shot[::-1] for shot in transformed_counts[i]
                 ]
+            if input_bit_circuit:
+                for input_bit_op in input_bit_circuit:
+                    for bit_result in regs.registers.bit_register_dict[input_bit_op.name()]:
+                        bit_result[input_bit_op.index()] = input_bit_op.value()
     elif sim_type == "statevector":
         for i, regs in enumerate(output_registers):
             vector = list(np.asarray(result.data(i)["statevector"]).flatten())
@@ -130,14 +140,15 @@ def _transform_job_result(
     sim_type: str,
     result: Result,
     output_registers: Union[RegistersWithLengths, List[RegistersWithLengths]],
+    input_bit_circuits: Union[Optional[Circuit], List[Optional[Circuit]]],
     res_index: Optional[int] = 0,
 ) -> Tuple[
     Dict[str, List[List[bool]]],
     Dict[str, List[List[float]]],
     Dict[str, List[List[complex]]],
 ]:
-    if isinstance(output_registers, list):
-        _transform_job_result_list(memory, sim_type, result, output_registers)
+    if isinstance(output_registers, list) and isinstance(input_bit_circuits, list):
+        _transform_job_result_list(memory, sim_type, result, output_registers, input_bit_circuits)
         final_output = RegistersWithLengths()
         for regs in output_registers:
             for key, value_bools in regs.registers.bit_register_dict.items():
@@ -156,6 +167,17 @@ def _transform_job_result(
                 else:
                     final_output.registers.complex_register_dict[key] = value_complexes
         return astuple(final_output.registers)
-    else:
-        _transform_job_result_single(memory, sim_type, result, output_registers, res_index)
+    elif isinstance(output_registers, RegistersWithLengths) and (
+        isinstance(input_bit_circuits, Circuit) or input_bit_circuits is None
+    ):
+        _transform_job_result_single(
+            memory,
+            sim_type,
+            result,
+            output_registers,
+            input_bit_circuits,
+            res_index,
+        )
         return astuple(output_registers.registers)
+    else:
+        raise ValueError("Invalid input for output_registers and/or input_bit_circuits.")